Check for missing values when counting reads (#380)
Read coordinates are converted to bounding boxes based on defined
exons. If a read starts before the first exon or ends after the
last exon in a gene then that coordinate is marked as -1. Novel
splice sites are also marked as -1. When checking if a read
should count toward an isoform the -1 values should be checked